move methods from eum_base to _geos_area
